Add files via upload
Extended the overall structure of the program. Revised, removed, added, and modified previous parts of the program as planning changed. Most notable change from last upload is the change in structure of the 'Acc_Create' function for creating accounts, from a single internalized function, into separate segmented functions, that run off of a primary one; which allows for much more flexibility for error-correcting, moving back to a previous section in the interface, and much more comprehensive, easy to understand, code. Future uploads within 24h will include the addition of TO-DO's for separate sections of the program, as well as cleanup from notes of debug code.